[DOP-4033]: Move commands into shared directory
