+++ src/commands/index.ts
@@ -1,3 +1,3 @@
-import { nextGenParse } from './src/next-gen-parse';
+import { nextGenParse } from './src/shared/next-gen-parse';
 
 export { nextGenParse };
