+++ src/commands/src/shared/next-gen-html.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import { promisify } from 'util';
-import { getRepoDir } from './helpers';
+import { getRepoDir } from '../helpers';
 
 const writeFileAsync = promisify(fs.writeFile);
